+++ tests/incremental_e2e.rs
@@ -42,7 +42,7 @@ fn retitling_invalidates_taxonomy_listing() {
     let site = Site::with(CONFIG);
     site.write(
         "config.kdl",
-        "site \"T\"\npaths {\n  content \"content\"\n  dist \"public\"\n}\n\\\ncontent {\n  taxonomies {\n    tags listing=#true\n  }\n}\n",
+        "site \"T\"\npaths {\n  content \"content\"\n  dist \"public\"\n}\n\\\ncontent {\n  taxonomies {\n    tags { listing }\n  }\n}\n",
     );
     site.write(
         "content/posts/a.typ",
@@ -711,7 +711,7 @@ fn renamed_page_prunes_the_old_permalink() {
 
 #[test]
 fn dropped_taxonomy_term_prunes_its_index() {
-    let config = format!("{CONFIG}content {{\n  taxonomies {{\n    tags listing=#true\n  }}\n}}\n");
+    let config = format!("{CONFIG}content {{\n  taxonomies {{\n    tags {{ listing }}\n  }}\n}}\n");
     let site = Site::with(&config);
     site.write(
         "content/a.typ",